slightly refactored TezosDatabaseOperations - moved helper functions to the util
