+++ src/main/scala/tech/cryptonomic/conseil/util/DatabaseUtil.scala
@@ -2,8 +2,8 @@ package tech.cryptonomic.conseil.util
 
 import slick.jdbc.PostgresProfile.api._
 import slick.jdbc.{GetResult, PositionedParameters, SQLActionBuilder}
-import tech.cryptonomic.conseil.generic.chain.DataTypes.{Predicate, QueryOrdering}
-import tech.cryptonomic.conseil.tezos.TezosDatabaseOperations.{makePredicates, makeOrdering, makeLimit}
+import tech.cryptonomic.conseil.generic.chain.DataTypes.OperationType.OperationType
+import tech.cryptonomic.conseil.generic.chain.DataTypes.{OperationType, Predicate, QueryOrdering}
 
 /**
   * Utility functions and members for common database operations.
@@ -11,76 +11,145 @@ import tech.cryptonomic.conseil.tezos.TezosDatabaseOperations.{makePredicates, m
 object DatabaseUtil {
   lazy val db = Database.forConfig("conseildb")
 
-  /** Concatenates SQLActionsBuilders
-    * Slick does not support easy concatenation of actions so we need this function based on https://github.com/slick/slick/issues/1161
-    * @param acc      base SqlActionBuilder with which we want concatenate other actions
-    * @param actions  list of actions to concatenate
-    * @return         one SQLActionBuilder containing concatenated actions
+  /**
+    * Utility object for generic query composition with SQL interpolation
     */
-  def concatenateSqlActions(acc: SQLActionBuilder, actions: List[SQLActionBuilder]): SQLActionBuilder = {
-    actions.foldLeft(acc) {
-      case (accumulator, action) =>
-        SQLActionBuilder(accumulator.queryParts ++ action.queryParts, (p: Unit, pp: PositionedParameters) => {
-          accumulator.unitPConv.apply(p, pp)
-          action.unitPConv.apply(p, pp)
-        })
-    }
-  }
-
-  /** Creates SQLAction of sequence of values
-    * @param  xs  list of values to be inserted into SQLAction
-    * @return     SqlActionBuilder with values from parameter
-    */
-  def insertValuesIntoSqlAction[T](xs: Seq[T]): SQLActionBuilder = {
-    var b = sql"("
-    var first = true
-    xs.foreach { x =>
-      if(first) first = false
-      else b = concatenateSqlActions(b, List(sql","))
-      b = concatenateSqlActions(b, List(sql"'#$x'"))
-    }
-    concatenateSqlActions(b, List(sql")"))
-  }
-
-  /** Implicit value that allows getting table row as Map[String, Any] */
-  implicit val getMap: GetResult[Map[String, Any]] = GetResult[Map[String, Any]](positionedResult => {
-    val metadata = positionedResult.rs.getMetaData
-    (1 to positionedResult.numColumns).flatMap(i => {
-      val columnName = metadata.getColumnName(i).toLowerCase
-      val columnValue = positionedResult.nextObjectOption
-      columnValue.map(columnName -> _)
-    }).toMap
-  })
-
-  /** Implicit class providing helper methods for SQLActionBuilder */
-  implicit class SqlActionHelper(action: SQLActionBuilder) {
-    /** Method for adding predicates to existing SQLAction
-      * @param predicates list of predicates to add
-      * @return           new SQLActionBuilder containing given predicates
+  object QueryBuilder {
+    /** Concatenates SQLActionsBuilders
+      * Slick does not support easy concatenation of actions so we need this function based on https://github.com/slick/slick/issues/1161
+      *
+      * @param acc     base SqlActionBuilder with which we want concatenate other actions
+      * @param actions list of actions to concatenate
+      * @return one SQLActionBuilder containing concatenated actions
       */
-    def addPredicates(predicates: List[Predicate]): SQLActionBuilder = {
-      concatenateSqlActions(action, makePredicates(predicates))
-    }
-
-    /** Method for adding ordering to existing SQLAction
-      * @param ordering   list of QueryOrdering to add
-      * @return           new SQLActionBuilder containing ordering statements
-      */
-    def addOrdering(ordering: List[QueryOrdering]): SQLActionBuilder = {
-      val queryOrdering = if(ordering.isEmpty) {
-        List.empty
-      } else {
-        List(makeOrdering(ordering))
+    def concatenateSqlActions(acc: SQLActionBuilder, actions: List[SQLActionBuilder]): SQLActionBuilder = {
+      actions.foldLeft(acc) {
+        case (accumulator, action) =>
+          SQLActionBuilder(accumulator.queryParts ++ action.queryParts, (p: Unit, pp: PositionedParameters) => {
+            accumulator.unitPConv.apply(p, pp)
+            action.unitPConv.apply(p, pp)
+          })
       }
-      concatenateSqlActions(action, queryOrdering)
     }
 
-    /** Method for adding limit to existing SQLAction
-      * @param limit      limit to add
-      * @return           new SQLActionBuilder containing limit statement
+    /** Creates SQLAction of sequence of values
+      *
+      * @param  xs list of values to be inserted into SQLAction
+      * @return SqlActionBuilder with values from parameter
       */
-    def addLimit(limit: Int): SQLActionBuilder = {
-      concatenateSqlActions(action, List(makeLimit(limit)))
+    def insertValuesIntoSqlAction[T](xs: Seq[T]): SQLActionBuilder = {
+      var b = sql"("
+      var first = true
+      xs.foreach { x =>
+        if (first) first = false
+        else b = concatenateSqlActions(b, List(sql","))
+        b = concatenateSqlActions(b, List(sql"'#$x'"))
+      }
+      concatenateSqlActions(b, List(sql")"))
+    }
+
+    /** Implicit value that allows getting table row as Map[String, Any] */
+    implicit val getMap: GetResult[Map[String, Any]] = GetResult[Map[String, Any]](positionedResult => {
+      val metadata = positionedResult.rs.getMetaData
+      (1 to positionedResult.numColumns).flatMap(i => {
+        val columnName = metadata.getColumnName(i).toLowerCase
+        val columnValue = positionedResult.nextObjectOption
+        columnValue.map(columnName -> _)
+      }).toMap
+    })
+
+    /** Implicit class providing helper methods for SQLActionBuilder */
+    implicit class SqlActionHelper(action: SQLActionBuilder) {
+      /** Method for adding predicates to existing SQLAction
+        *
+        * @param predicates list of predicates to add
+        * @return new SQLActionBuilder containing given predicates
+        */
+      def addPredicates(predicates: List[Predicate]): SQLActionBuilder = {
+        concatenateSqlActions(action, makePredicates(predicates))
+      }
+
+      /** Method for adding ordering to existing SQLAction
+        *
+        * @param ordering list of QueryOrdering to add
+        * @return new SQLActionBuilder containing ordering statements
+        */
+      def addOrdering(ordering: List[QueryOrdering]): SQLActionBuilder = {
+        val queryOrdering = if (ordering.isEmpty) {
+          List.empty
+        } else {
+          List(makeOrdering(ordering))
+        }
+        concatenateSqlActions(action, queryOrdering)
+      }
+
+      /** Method for adding limit to existing SQLAction
+        *
+        * @param limit limit to add
+        * @return new SQLActionBuilder containing limit statement
+        */
+      def addLimit(limit: Int): SQLActionBuilder = {
+        concatenateSqlActions(action, List(makeLimit(limit)))
+      }
+    }
+
+    /** Prepares predicates and transforms them into SQLActionBuilders
+      *
+      * @param  predicates list of predicates to be transformed
+      * @return list of transformed predicates
+      */
+    def makePredicates(predicates: List[Predicate]): List[SQLActionBuilder] =
+      predicates.map { predicate =>
+        concatenateSqlActions(
+          predicate.precision.map(precision => sql""" AND ROUND(#${predicate.field}, $precision) """)
+            .getOrElse(sql""" AND #${predicate.field} """),
+          List(mapOperationToSQL(predicate.operation, predicate.inverse, predicate.set.map(_.toString)))
+        )
+      }
+
+    /** Prepares query
+      *
+      * @param table   table on which query will be executed
+      * @param columns columns which are selected from teh table
+      * @return SQLAction with basic query
+      */
+    def makeQuery(table: String, columns: List[String]): SQLActionBuilder = {
+      val cols = if (columns.isEmpty) "*" else columns.mkString(",")
+      sql"""SELECT #$cols FROM #$table WHERE true """
+    }
+
+    /** Prepares ordering parameters
+      *
+      * @param ordering list of ordering parameters
+      * @return SQLAction with ordering
+      */
+    def makeOrdering(ordering: List[QueryOrdering]): SQLActionBuilder = {
+      val orderingBy = ordering.map(x => s"${x.field} ${x.direction}").mkString(",")
+      sql""" ORDER BY #$orderingBy"""
+    }
+
+    /** Prepares limit parameters
+      *
+      * @param limit list of ordering parameters
+      * @return SQLAction with ordering
+      */
+    def makeLimit(limit: Int): SQLActionBuilder = {
+      sql""" LIMIT $limit"""
+    }
+
+    /** maps operation type to SQL operation */
+    private def mapOperationToSQL(operation: OperationType, inverse: Boolean, vals: List[String]): SQLActionBuilder = {
+      val op = operation match {
+        case OperationType.between => sql"BETWEEN #${vals.head} AND #${vals(1)}"
+        case OperationType.in => concatenateSqlActions(sql"IN ", List(insertValuesIntoSqlAction(vals)))
+        case OperationType.like => sql"LIKE '%#${vals.head}%'"
+        case OperationType.lt | OperationType.before => sql"< '#${vals.head}'"
+        case OperationType.gt | OperationType.after => sql"> '#${vals.head}'"
+        case OperationType.eq => sql"= '#${vals.head}'"
+        case OperationType.startsWith => sql"LIKE '#${vals.head}%'"
+        case OperationType.endsWith => sql"LIKE '%#${vals.head}'"
+      }
+      concatenateSqlActions(op, List(sql" IS #${!inverse}"))
     }
   }
